requirements.txt
mirobot-py
wlkata-mirobot-python
asyncua
numpy
black